Publish PubSub token to notify initialization of GameViewDom div in initPage()
The actual Pubsub token is pubSubTokens.gameViewInitialized taken from
pubSubTokens module.

On the other hand, on GameController class, and in particular
in #initGameView method, a subscription to this same token is done.

When the GameViewDom instance's div is actually initialized, this
ensures that the #makeFirstMove method is called, which is in charge of
determining the first player to make a move, and that make it play the
move (although this last functionality is not implemented yet).

Note that #makeFirstMove also unsubscribes from the
pubSubTokens.gameViewInitialized token which is used to trigger its
execution.