Add failing test for Gameboard class: make receiveAttack() method return a outcome code
Before this, that method returned true or false in case of hit or miss,
respectively.

Now, make it return an outcome code, to report the case of sink, too:
0: miss
1: hit
2: hit and sink

The previous tests checking the return values to be truthy or falsy
still pass, since hit and hit and sink include both the hit of a ship.