Make receiveAttack() method of gameController return an outcome code to make the previous test pass
Before this, the method returned true or false in case of hit or miss,
respectively.

Now, make it returns an outcome code, to report the case of sink, too:
0: miss
1: hit
2: hit and sink

Note that the cases of 'hit' and 'hit and sink' both include the case of
hitting a ship, and the corresponding outcomeCode is truthy, while this
is falsy otherwise.