Refactor #playGame() method of GameController class to trigger new turns via PubSub
Before this, there was an infinite while loop that was interrupted when
the a player won the game. We need a different mechanism because the
next functionalities will be triggered via PubSub.

Now, #playGame() method subscribes to pubSubTokens.playTurn token
(defined in pubSubTokens module) which triggers the execution of a new
turn, calling #playTurn() when the token is captured.

In #playTurn(), when the turn ends, a pubSubTokens.playTurn is published
to start a new turn. When a player wins, it unsubscribes from the
pubSubTokens.playTurn token.