Refactor GameController module to handle attack coords acquisition via PubSub
In #playTurn() method, we first subscribe to some
pubSubTokens.attackCoordsAcquired token (defined in pubSubTokens module)
to specify through #attackCoordsAcquiredCallback what we want to do once
the coordinates to attack are acquired (ie, attacking the opponent, and
so on), and then trigger the acquisiton of such coords. Note that all
the remaining code of this method has been moved in the callback
 #attackCoordsAcquiredCallback: for the human player, the acquisition of
the coords (not implemented yet) is asynchronous, so this helps handling
it.

Note that in #getAttackCoords a pubSubTokens.attackCoordsAcquired token
is published with the acquired coords to attack. As already said, the
human player is not implemented yet, so a placeholder coordinate is
used, which raises an error in the console (due to attempting to attack
the same cell twice). This will be fixed next.