Implement functionality to show attack outcome on opponent gameboard
In #attackCoordsAcquiredCallback() method of GameController class, we
first subscribe to pubSubTokens.attackOutcomeShown token (defined in
pubSubTokens.js), which will be used to notify when the UI has been
updated to be able to proceed handling the game turn.

Then, the opponent is attacked and throught the #showAttackOutcome
method a pubSubTokensUi.showAttackOutcome(this.#opponent) token is
published to trigger the update of the UI.

So far, a simple cell background color is used to show whether the
attack was a hit or a miss. In the future, icons might be shown, or even
sounds played.