Add PlaceableObjectDom module to place object on grid containers
This module defines the PlaceableObjectDom class which initializes a div
that setup the occupied row/columns grid element and other styles to
make them occupy the coordinates specified in the constructor.

This will be used for placing ships, and hit/miss marks on the
gameboard. In particular, classes extending this class will be created,
to add specific properties.