Add getShipPosition method to Gameboard class to make the previous test pass
A private property #fleetPosition is added, which is a Map object
including an entry for each ship in the fleet, indexed by ship name.
For not deployed ships the value is null, while for the deployed / sunk
ships it is an array: [occupiedCellsCoordsArray,direction].

The getShipPosition has the ship name as parameter and returns the
corresponding value in #fleetPosition.