Add hideDeployedFleet() method to GameboardDom class
The private property #deployedFleetDom has been added to teh
GameboardDom class, which has the ShipDom object for each deployed ship.

The #showShip() method has been refactored to just append a shipDomObj
to the gameboard div, while the #hideShip() method removes it from the
same div. The #createShipDom() method just creates a new ShipDom object.

The showDeployedFleet() method, for each ship from
this.#gameboard.deployedFleet, either creates and saves a ShipDom object
in #deployedFleetDom (of not present in there) or retrieves it from
there and then calls the  #showShip() method to display it.

Then, the hideDeployedFleet() method calls the #hideShip() method for
all ShipDom objects in #deployedFleetDom.