Add PlayersNameViewDom module
The module defines the PlayersNameViewDom class which initialize a div
that can be used to select the players'names. This depends whether there
are one or two players. In the former case, the opponent is the AI.

HomeViewDom has been edited to be able to show the correct
PlayersNameViewDom view.

Clicking on the .play-btn initialize a new game with the correct number
of players.

No style is applied, yet, nor a new game is initialized.