Refactor AiPlayer class to use a Map internally to keep track of possible targets
The map is indexed by string coordsArray.join(',') and stores the
coordsArray=[column, row] as a value. The reason is that if coordsArray
is not the one used as key (eg, a copy), it cannot be used to get the
corresponding data (while this is possible when that is transformed to
a string).

Manipulating the coordinates array to be used as indices will be used by
imporved versions of Ai players.