Add failing test: an AI player can apply the HuntTarget strategy
The strategy is the following:

There is a global list of possible targets. When a cell is attacked, it
is removed from such list.

Then, chose a random target as a base. However, when you score a hit,
add the neighbouring cells which have not been attacked yet to some high
priority list. When the high priority list is not empty, chose the next
cell to attack among these instead of the global one.

See https://www.datagenetics.com/blog/december32011/
See https://towardsdatascience.com/coding-an-intelligent-battleship-agent-bf0064a4b319