Improve the improvedHuntTarget strategy of AI Players: autoupdate opponentMinShipSize
When an oponent ship is sunk, the opponentMinShipSize is updated.

In particular, a list of opponents's  unsunk ship length is kept. Note
that this list is initialized with the ships' length of the player
itself, as both the player and the opponent have the same kind of ships
deployed. When an opponent ship is sunk, it is removed from such list,
and the opponentMinShipSize is updated.