Add click event listener to rotate ships when deploying the fleet
The PlayerDom class and GameboardDom class are being given a
canBeModified argument in the constructor, which defaults to false.

The effect of this argument, when true, is to initialize some callbacks
in the gameboard div to let the user modifiy it. In particular,
exploiting event delegation, a click event listener is added to the
gameboard div to be able to rotate a ship (if present) around the
clicked cell.

Note, however, that so far only a message is shown on the console, and
thus the corresponding ship is not actually rotated yet.

The canBeModified parameter is set to true only for the PlayerDom (and
then GameboardDom) object created in the DeployFleetViewDom class.