Implement drag handling of ships' divs in gameboardDom (in edit mode)
This has been implemented based on: https://javascript.info/mouse-drag-and-drop

When dragging, the ship div position is updated by applying a css
transform with translate (other than the existing rotate one, if
present), and not by absolutely positioning the element like in the
reference above.

However, right now, when releasing the drag, the ship returns to the
original position: the actual move to the new cells is not implemented,
yet.

The functionality should work both for descktop and mobile, for both
mouse and touch pointers. Indeed, pointerDown, pointerMove, and
pointerUp event handlers have been set.

Even in this case, a single event listener is used in the gameboard div:
event delegation is used to select the correct ship div.

Note however, that currently, the dragging conflicts with the click
event handling rotating the ships. Sometimes, when releasing the
dragging, the rotation is still performed, even if it shouldn't (eg,
when the pointer is a mouse, and release the ship in the initial
position). This will be fixed in a future commit.