Add getNestedElementOfClass(point, requiredClass) utility function
An helper function getNestedElementOfClass(point, requiredClass) has
been written to get the nested element at position given by
point = [x,y], which has the css class requiredClass.

This function uses the document.elementsFromPoint method, which returns
an array of all the nested element in a given point.

Then, this function has been used in the events listeners in
GameboardDom module, to get the required element (recall that event
listener in there are attached to the gameboard div, and exploit event
delegation).