Refactor edit event listeners in GameboardDom module to handle click to rotate and drag correctly
When setting up the edit mode in a GameboardDom object, a single event
listener to a pointerdown event is set (now called
 #startEditingPointerDownCallback).

In this event handler, after getting the pointed cellDiv and shipDiv, if
present (recall event delegation is used), a dragOn variable is
initialized to false, meaning that initially this event is assumed to be
part of a click event, which means: rotate the ship around this cell.

A pointermove event listener is then defined, which determines whether
we should initialize a drag, instead of rotating the ship. In
particular, a displacement threshold is defined: within a given initial
displacement, the event is still considered a click. Beyond that, a drag
is to be performed and the dragOn is set to true. As the dragOn is true,
the ship div is opportunely translated based on the pointer position.

See: https://stackoverflow.com/questions/6042202/how-to-distinguish-mouse-click-and-drag

Finally, a pointerup event listener is also defined, which based on the
dragOn select the correct action to perform (rotation or move).

Note however that the move div ship is not implemented, yet.