Enable the 'move a ship' functionality in GameboardDom class
The Gameboard class method startMoveShip() is called when the drags
begins, in the onPointerMoveCallback() event handler.

On the stopEditingPointerUpCallback event handler, if there was a drag
that has just ended, the cell to (try to) move to is determined: it is
either the cell in which the pointer has been released, or the original
cell if that does not exists.

Then, the Gameboard class method endMoveShip() is called, and after
that, the actual ship position is opportunely set.

Note: there are some spurious animations when releasing the ship in a
new location. This will be fixed later on.