Fix the spurious animations when dragging and releasing the ship in a new location
During the dragging, the ship position was free, while when releasing it
in a new location, it is constrained to the grid.

To allow for a smooth transition, first translate the ship in the new
position, and wait for the animation to end.

We now want to actually move the ship in the new location (in fact, it
is still placed in the original position..just translated). Since the
css transition property could be non-zero, force it to be zero, modify
the ship position, and restore the original ship css transform propery.

Then, restore the css transition property, too, after some artificial
delay to ensure the previous operations were performed.