Change the cursor when dragging a ship in a cell where it cannot be placed
While dragging a ship in the gameboard div, toggle the
'forbidden-position' css class in the ship div if the corresponding ship
cannot be placed in the pointed cell.

This is used to change the cursor to 'not-allowed' when
'forbidden-position' class is applied, so the user can be aware of that.