Add messages to GameViewDom view to tell each move outcome
The .msgP p element is updated via PubSub to show the attack outcome
message (for hit, miss, hit and sunk), as well as the current player to
play. To do that, a pubSubTokensUi.setGameStatusMsg token is published.

The message depends on whether there is 1 or 2 players, and whether the
first player is the ai or not.

Note, however, that showing a message necessaily slows down the game, as
otherwise it would not be readible. Some additional waiting time needs
to be inserted. Consider adding a button to disable such messages, and
make the gameplay faster.